Nearly Complete Report
Finished touching up some parts in the analysis and completed the report outline section. Moved the content descriptors from the table of contents into the report outline and summarized them to fit the word count. Just need to reduce some content to fully fit the word count, sort the in-text citations and remove any used ones from the bibliography.